zsh install & list

+++ zsh/install.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
 
-cd "$HOME"
-mkdir github
-cd github
+ 
+mkdir -p "$HOME"/github
+cd "$HOME"/github
 git clone https://github.com/robbyrussell/oh-my-zsh
-ln -s "$PWD"/oh-my-zsh "$HOME"/.oh-my-zsh
+[ ! -e "$HOME"/.oh-my-zsh ] && ln -s "$PWD"/oh-my-zsh "$HOME"/.oh-my-zsh
+cd -
+[ ! -e "$HOME"/.zshrc ] && cp -a "$HOME"/github/oh-my-zsh/templates/zshrc.zsh-template "$HOME"/.zshrc
 cp -ai dio.zsh-theme "$HOME"/.oh-my-zsh/themes/
 cp -ai dircolors.256dark "$HOME"/.dir_colors
 
+grep -q "^TERM=" "$HOME"/.zshrc || {
+  echo TERM=xterm-256color >> "$HOME"/.zshrc
+}
+
+grep -q "\.dir_colors" "$HOME"/.zshrc || {
+  echo eval '`dircolors $HOME/.dir_colors`' >> "$HOME"/.zshrc
+}
 echo 'OK. Now change change ZSH_THEME to "dio" in your .zshrc (ZSH_THEME="dio")'
-echo 'Also, make sure you have this in your .zshrc:
-TERM=xterm-256color
-`dircolors $HOME/.dir_colors`'
